// server/src/interfaces/Recipe.ts
import { ObjectId } from "mongodb";

export interface Recipe {
  id: string | ObjectId;
  title: string;
  description: string;
  categories: string[];
  ingredients: string[];
  steps: string[];
  createdAt: number;
  updatedAt?: number;
}

export interface RecipePayload extends Omit<Recipe, "id" | "updatedAt"> {}

export interface RecipePatchPayload extends Partial<RecipePayload> {}

export interface RecipeCreateRequestPayload
  extends Omit<Recipe, "id" | "updatedAt" | "createdAt"> {}